Ensure Ray cores >= BatchTokenizer cores
Tests in 5 files fail when PyTest tests are run on machines with
more than 10 cores. To fix this issue, Ray is initialized with more
cores to ensure adequate logical cpus are available. In addition,
one line of code causes failures on machines with more than 20 cores.
This issue was also fixed.