Ensure Ray cores >= BatchTokenizer cores and Levanter works with Pyenv (#260)
* Ensure Ray cores >= BatchTokenizer cores

Tests in 5 files fail when PyTest tests are run on machines with
more than 10 cores. To fix this issue, Ray is initialized with more
cores to ensure adequate logical cpus are available. In addition,
one line of code causes failures on machines with more than 20 cores.
This issue was also fixed.

* Fix src/levanter/logging.py for pyenv

Levanter works with pyenv with the exception of this file, which
hardcodes 'conda' and 'venv'. 'pyenv' is added to this hardcoded
array list to ensure pyenv works with Levanter. After making this
change, all PyTest tests pass and the sample comands in the
getting started tutorial work with pyenv.

---------

Co-authored-by: Sunil Pinnamaneni <sunil@exactnote.com>