A few cleanups for model instantiation.
Make the HF converter an instance method instead of a class property. This
makes it easier to instantiate a model family like LLama with a different
reference checkpoint.

Change the HF checkpoint converter to always accept an LmConfig instead of a
bare model or config object. There were only a few places where the model was
still being passed in.