Supporting varied mixtures over training (#868)
## Description
Currently, LM mixture dataset can only handle a static mixture over the
course of training. This PR enables varying this mixture over datasets
over the course of training. The user can now specify a list of stages
and the sequence index at which each should start.

Internally, we identify a training block to its stage, which defines its
mixing weights. To efficiently translate a data point's index within a
block to a respective source dataset, we precompute prefix sums that
track how many data points are seen by previous stages.

## Fixes Issues
https://github.com/stanford-crfm/marin/issues/81

## Unit test coverage
There are new unit tests in `test_varying_mixture.py` to ensure that the
varying mixture behaves as expected.

## Known breaking changes/behaviors
The design enables traditional usage of the MixtureDataset class.
However, some of the private quantities are different (i.e. the expected
counts per block now depends on the block and is not a member variable).
To my knowledge, these variables are not accessed outside of tests.

## Additional context
I have some changes I want to make Marin to enable usage of this new
functionality, though these updates are modular and can be seperate
PR's. I have spot-checked that training proceeds as expected with this
test. This is my first PR so feedback is appreciated :))