Initial implementation of power scheduler (#974)
Implementation based on this paper. https://arxiv.org/abs/2408.13359. 

@dlwh @Helw150 I'm not sure if this is the best way to go about
implementing this as is. I had to add five more parameters to the
parameter list of the OptimizerConfig class to implement Power
Scheduler, as, based on my understanding, there is a tight coupling
between the optimizer and the scheduler. Should there be an effort to
decouple the optimizer and the scheduler, so that we don't have to pass
schedule-specific parameters directly into OptimizerConfig?