Allow overriding `reference_checkpoint` when using HF Checkpoints (#969)
## Description

When loading checkpoints for models from HF, levanter expects a HF Hub
repo name/path as a reference. If there is some mismatch between the
default `reference` and the one which we wish to use, we have to call
`HFCheckpointConverter`'s `replaced` method which is unintuitive API.

This PR allows overriding the `reference_checkpoint` when invoking the
checkpoint converter method directly, thus making it easier for users to
point to different HF Hub checkpoints.

## Known breaking changes/behaviors
Non-breaking change as the arguments are `Optional` and revert to
defaults if not overriden.