Implement OpenAI compatible inference service. (#1159)
Features:

- OpenAI compatible service
- On demand checkpoint reloading

Current limitations:

- Multi-host support

Multiple hosts may or may not work. They'll definitely start their own
inference server port, to guarantee maximum chaos.

- Checkpoint sharding is not totally clear

This should be updated to use a global barrier to ensure workers are
aligned with the leader etc. This shouldn't be an issue however, as we
should always be using the lead node for submitting requests

- Model selection is ignored.
This is intentional, as we will in general be loading models explicitly
at startup or via a checkpoint service.

- No cancellation support.

This is a nice to have, for example if we want to cancel some queries
when loading new weights or if they are slow.

- Missing performance for multi-turn queries.
Some kind of persistent KV-caching will be necessary for multi-turn
(agent) RL testing. See
https://github.com/jax-ml/jax-llm-examples/blob/main/serving/serving_jax/serving_loop.py
for an example.